Time: 170 ms (5.81%), Space: 47.7 MB (6.52%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -13,7 +13,7 @@ class Solution {
                 int l = n-1;
                 
                 while(k<l){
-                    int sum = nums[i] + nums[j] + nums[k] + nums[l];
+                    long sum = (long)nums[i] + nums[j] + nums[k] + nums[l];
                     if(sum == target){
                         set.add(Arrays.asList(nums[i],nums[j], nums[k], nums[l]));
                         k++;
